A bit of order in this mess
