+++ tests/gui/test_cs_panel.py
@@ -5,7 +5,6 @@ from PyQt6 import QtWidgets
 
 from i19serial_ui.gui.widgets.cs_panel import (
     CoordinateSystemPanel,
-    _calculate_kapton_xz_positions,
 )
 from i19serial_ui.parameters.coordinates import Coord3D, Coordinates, FiducialPosition
 from i19serial_ui.parameters.grid import GridType
@@ -15,21 +14,6 @@ FAKE_COORDS = {
     "top_right": Coord3D(0.1, 0.0, 1.4),
     "bottom_left": Coord3D(0.3, 0.0, 1.2),
 }
-
-
-@pytest.mark.parametrize(
-    "fiducial, xz, expected_xz",
-    [
-        (FiducialPosition.TL, (0.2, 1.0), (0.350, 0.850)),
-        (FiducialPosition.TR, (0.2, 1.0), (0.050, 0.850)),
-        (FiducialPosition.BL, (0.2, 1.0), (0.350, 1.150)),
-    ],
-)
-def test_calculate_kapton_xz_positions(fiducial, xz, expected_xz):
-    res = _calculate_kapton_xz_positions(xz, fiducial)
-
-    assert res[0] == pytest.approx(expected_xz[0], 1e-3)
-    assert res[1] == pytest.approx(expected_xz[1], 1e-3)
 
 
 @pytest.fixture
